+++ magicctapipe/image/__init__.py
@@ -9,10 +9,17 @@ from .leakage import (
     get_leakage,
 )
 
+from .calib import (
+    Calibrate_LST, 
+    Calibrate_MAGIC
+)
+
 __all__ = [
     "MAGICClean",
     "PixelTreatment",
     "get_num_islands_MAGIC",
     "clean_image_params",
     "get_leakage",
+    "Calibrate_LST", 
+    "Calibrate_MAGIC"
 ]
